Comment out 'peak's Threebox layer
Comment out Threebox layer 'peaks', which creates the 3D heatmap; this is because the 3D heatmap is not perfect / presentable at the moment, and while we need to store its code, I want to keep it hidden in case this project needs to be demo'ed on short notice. The demo should not have half-finished parts in it.